Refactor key value list component
