+++ app/javascript/components/terraform-template-catalog-form/index.jsx
@@ -4,7 +4,6 @@ import PropTypes from 'prop-types';
 import createSchema from './terraform-template-catalog-form.schema';
 import componentMapper from '../../forms/mappers/componentMapper';
 import {
-  KeyValueListComponent,
   //   CopyFromProvisonButton,
   TreeViewReduxWrapper,
   conditionalCheckbox,
@@ -143,7 +142,6 @@ const TerraformTemplateCatalogForm = ({ initialData }) => {
 
   const mapper = {
     ...componentMapper,
-    'key-value-list': KeyValueListComponent,
     // 'copy-from-provisioning': CopyFromProvisonButton,
     'tree-view-redux': TreeViewReduxWrapper,
     'conditional-checkbox': conditionalCheckbox,
